feat(vercel-ai-sdk): add media processing
