+++ package.json
@@ -28,8 +28,8 @@
   "author": "Langfuse",
   "license": "MIT",
   "devDependencies": {
-    "@ai-sdk/anthropic": "^1",
-    "@ai-sdk/openai": "^1",
+    "@ai-sdk/anthropic": "^2",
+    "@ai-sdk/openai": "^2",
     "@langchain/core": "^0.3.66",
     "@langchain/langgraph": "^0.4.1",
     "@langchain/openai": "^0.6.3",
@@ -45,7 +45,7 @@
     "@types/mustache": "^4.2.6",
     "@typescript-eslint/eslint-plugin": "^8.36.0",
     "@typescript-eslint/parser": "^8.39.0",
-    "ai": "^4",
+    "ai": "^5",
     "dotenv": "^17.2.0",
     "eslint": "^9.32.0",
     "eslint-config-prettier": "^10.1.8",
